Update priority-override-for-trigger-campaigns.md
Proposing change after asking Marketo Support for clarification (case 02635388). On 4/20/2023, Support said: "The limit is in reference to increasing the trigger campaign priority only (e.g. changing it from medium to high), and there is no limitation for lowering the trigger campaign priority (e.g. changing it from high to medium)." The verbiage here is imprecise and it would be improved by clarifying.